moved Examples file to frameworks package Deleted old API implementation
