+++ src/main/java/frameworks/OpenAITextOrganizer.java
@@ -1,6 +1,5 @@
 package frameworks;
 
-import api.Examples;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
